Split out savePost
